temporal: SOOO Slow operations

+++ RealtimeBackend/src/environment/spawns/spawn.rs
@@ -3,5 +3,5 @@ use async_trait::async_trait;
 #[async_trait]
 pub trait Spawn {
     async fn run(&self);
-    async fn spawn_enemies(&self, amount: u8);
+    async fn spawn_enemies(&self, amount: usize);
 }
